requirements.txt
google-generativeai==0.8.0
transformers==4.46.3
pyyaml==6.0.3
torch==2.10.0+cpu
